Addition of login page and the user can now report a bug as not being fixed, sql command in text box bug also has be en fixed.

+++ Debugger/Developer.cs
@@ -12,23 +12,27 @@ using ColorCode;
 
 namespace Debugger
 {
+    /// <summary>
+    ///  Contains methods for viewing, deleting, editing source code
+    ///  and archiving them.
+    /// </summary>
     public partial class Developer : Form
     {
         public Developer()
         {
             InitializeComponent();
-            txtDate.ReadOnly = true;
+            txtDate.ReadOnly = true;                                    
             txtDate.Text = DateTime.Now.ToString("dd-mm-yyyy");          // Automaticly genarate the date
             displayBugs();
         }
 
         /// <summary>
-        /// Display the data in the BugTable  to the develop in a data grid view
+        /// Display the data in the BugTable to the develop in a data grid view
         /// </summary>
         public void displayBugs()
         {
 
-            String sqlCon = (@"Data Source=(LocalDB)\MSSQLLocalDB;AttachDbFilename=F:\vs\Debugger\BugDatabase.mdf;Integrated Security=True;Connect Timeout=30");   // display all data in the datbase in a data grid view when the displayed button is pressed 
+            String sqlCon = (@"Data Source=(LocalDB)\MSSQLLocalDB;AttachDbFilename=F:\vs\Debugger\BugDatabase.mdf;Integrated Security=True;Connect Timeout=30");   // display all data in the datbase in a data grid view when the displayed method is called 
             SqlConnection con = new SqlConnection(sqlCon);
             con.Open();
             DataTable dt = new DataTable();
@@ -36,7 +40,7 @@ namespace Debugger
             adapt.Fill(dt);
             dataGridViewDeveloper.DataSource = dt;
 
-            this.dataGridViewDeveloper.Columns["Name"].Visible = false;                                                                                                       // hide the three columns that the white box doesnt need to see
+            this.dataGridViewDeveloper.Columns["Name"].Visible = false;                                                                                                       // Hide the three columns that the white box tester doesnt need to see
             this.dataGridViewDeveloper.Columns["Date "].Visible = false;
             this.dataGridViewDeveloper.Columns["Comment"].Visible = false;
             con.Close();
@@ -47,14 +51,19 @@ namespace Debugger
         /// </summary>
         /// <param name="sender"></param>
         /// <param name="e"></param>
-        private void btnBackPro_Click(object sender, EventArgs e)         // switch back too main menu
+        private void btnBackPro_Click(object sender, EventArgs e)        
         {
             this.Hide();
             mainMenu main = new mainMenu();
             main.ShowDialog();
         }
 
-        private void Programmer_FormClosed(object sender, FormClosedEventArgs e)  // switch back to the main menu
+        /// <summary>
+        /// Exit the entire aplication
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
+        private void Programmer_FormClosed(object sender, FormClosedEventArgs e)  
         {
             Application.Exit();
         }
@@ -63,20 +72,20 @@ namespace Debugger
         /// </summary>
         /// <param name="sender"></param>
         /// <param name="e"></param>
-        private void dataGridView1_CellClick(object sender, DataGridViewCellEventArgs e)                                 //display the bug in formation in the list view when selected
+        private void dataGridView1_CellClick(object sender, DataGridViewCellEventArgs e)                 
         {
             lsBug.Items.Clear();
-            if (e.RowIndex != -1)
+            if (e.RowIndex != -1)                                                                                                                           // Stop the user form selecting the menu bar.
             {
-               dataGridViewDeveloper.ReadOnly = true;
+               dataGridViewDeveloper.ReadOnly = true;                                                                                                       // // Stop  the user entering data to the data grid view.
 
                 String sourceCode = dataGridViewDeveloper.Rows[e.RowIndex].Cells[11].Value.ToString();
                 string colourSourceCode = new CodeColorizer().Colorize(sourceCode, Languages.CSharp);
                 string html = ("<!doctype html><head><meta charset=\"utf-8\" <title> Code Snippet </title> </head> <body>" + colourSourceCode + "</body></html>");
                 webColourCode.DocumentText = html;
 
-                txtBugID.Text = dataGridViewDeveloper.Rows[e.RowIndex].Cells[0].Value.ToString();
-  
+                txtBugID.Text = dataGridViewDeveloper.Rows[e.RowIndex].Cells[0].Value.ToString();                                                          // Display information about the bug from the clicked cell. 
+
                 lsBug.Items.Add(("Bug ID:                 ")+dataGridViewDeveloper.Rows[e.RowIndex].Cells[0].Value.ToString());
                 lsBug.Items.Add(("ApplicationName: ") + dataGridViewDeveloper.Rows[e.RowIndex].Cells[1].Value.ToString());
                 lsBug.Items.Add(("Bug Stmptoms:     ") + dataGridViewDeveloper.Rows[e.RowIndex].Cells[2].Value.ToString());
@@ -99,7 +108,8 @@ namespace Debugger
 
       
         /// <summary>
-        ///  Search and return bug information from the table to the developer
+        ///  Search and return bug information from the table to 
+        ///  the developer display the information in the corresponding feilds
         /// </summary>
         /// <param name="sender"></param>
         /// <param name="e"></param>
@@ -111,7 +121,7 @@ namespace Debugger
             SqlCommand selcmd = new SqlCommand("SELECT BugID, ApplicationName, BugSymptoms,BugTrigger,ProjectName,SourceFile,ClassName,LineNumber,Method,CodeBlock, CodeAuthor, SourceCode, Name, Date, Comment FROM BugTable WHERE BugID ='" + txtBugID.Text + "'", con);   // select all bug information on the row form bug id
             SqlDataReader mySqlDataReader = selcmd.ExecuteReader();
 
-            while (mySqlDataReader.Read())
+            while (mySqlDataReader.Read())                                                                                                                                               // While there is data in the row display the data to the user
             {
                 lsBug.Items.Clear();
 
@@ -126,16 +136,42 @@ namespace Debugger
                 lsBug.Items.Add(("Method:                  ") + mySqlDataReader["Method"].ToString());
                 lsBug.Items.Add(("CodeBlock:            ") + mySqlDataReader["CodeBlock"].ToString());
                 lsBug.Items.Add(("Code Author:          ") + mySqlDataReader["CodeAuthor"].ToString());
-                txtSourceCode.Text = mySqlDataReader["SourceCode"].ToString();
+
+                txtSourceCode.Text = mySqlDataReader["SourceCode"].ToString();                                                                                    
                 String sourceCode = mySqlDataReader["SourceCode"].ToString();
-                string colourSourceCode = new CodeColorizer().Colorize(sourceCode, Languages.CSharp);
+                string colourSourceCode = new CodeColorizer().Colorize(sourceCode, Languages.CSharp);                                                                                   // Format Source code with colour for c# code with the colour code libary
                 string html = ("<!doctype html><head><meta charset=\"utf-8\" <title> Code Snippet </title> </head> <body>" + colourSourceCode + "</body></html>");
                 webColourCode.DocumentText = html;
             }
             con.Close();
         }
+
         /// <summary>
-        /// The developer can archive bugs when  they arefixed to store them in a new table and delete them form the bug Table  
+        /// Use parameters to update the table when called
+        /// </summary>
+        /// <param name="name"></param>
+        /// <param name="date"></param>
+        /// <param name="comment"></param>
+        /// <param name="sourceCode"></param>
+        /// <param name="commandString"></param>
+        public void UpdateBug(String name, String date, String comment, String sourceCode,  String commandString)
+        {
+            String sqlCon = (@"Data Source=(LocalDB)\MSSQLLocalDB;AttachDbFilename=F:\vs\Debugger\BugDatabase.mdf;Integrated Security=True;Connect Timeout=30");
+            SqlConnection con = new SqlConnection(sqlCon);
+            con.Open();
+            SqlCommand updateCommand = new SqlCommand(commandString, con);
+
+            updateCommand.Parameters.AddWithValue("@name", name);
+            updateCommand.Parameters.AddWithValue("@date", date);
+            updateCommand.Parameters.AddWithValue("@comment", comment);
+            updateCommand.Parameters.AddWithValue("@sourceCode", sourceCode);
+            updateCommand.ExecuteNonQuery();
+            con.Close();
+           
+        }
+        /// <summary>
+        /// The developer can archive bugs when they are fixed to store 
+        /// them in a new table and delete them form the bug table  
         /// </summary>
         /// <param name="sender"></param>
         /// <param name="e"></param>
@@ -149,28 +185,33 @@ namespace Debugger
             String sourceCode = txtSourceCode.Text;
             String sqlCon = (@"Data Source=(LocalDB)\MSSQLLocalDB;AttachDbFilename=F:\vs\Debugger\BugDatabase.mdf;Integrated Security=True;Connect Timeout=30");
 
-            if (bugID != "")
+            if (bugID != "" && name !=""&& comment!="")                                                                                                                  //Check if all fields have been filled if not notify user
             {
                 SqlConnection con = new SqlConnection(sqlCon);
-                SqlCommand upadate = new SqlCommand("UPDATE BugTable SET Name ='" + name + "', Date ='" + date + "', comment ='" + comment + "', SourceCode ='" + sourceCode  + "' WHERE (BugID ='" + bugID + "')", con);
-                SqlCommand archive = new SqlCommand("INSERT INTO ArchivedBugsTable (ApplicationName, BugSymptoms, BugTrigger, ProjectName, SourceFile, ClassName, LineNumber, Method, CodeBlock, CodeAuthor, SourceCode, Name, Date, Comment)"
-                    + " SELECT ApplicationName, BugSymptoms, BugTrigger, ProjectName, SourceFile, ClassName, LineNumber, Method, CodeBlock, CodeAuthor, SourceCode, Name, Date, Comment FROM BugTable WHERE BugID ='" + txtBugID.Text + "'", con);
+                String commandString = ("UPDATE BugTable SET Name = @name, Date =@date, comment =@comment, SourceCode = @sourceCode WHERE (BugID ='" + bugID + "')");    // Give the conecetion string the update query 
+                UpdateBug(txtName.Text,txtDate.Text,txtComent.Text,txtSourceCode.Text,commandString);                                                                    // Pass the update method the values to be updated from the user                   
+                SqlCommand archive = new SqlCommand("INSERT INTO ArchivedBugsTable (ApplicationName, BugSymptoms, BugTrigger, ProjectName, SourceFile, ClassName, LineNumber, Method, CodeBlock, CodeAuthor, SourceCode, Name, Date, Comment)"      // Bug has been updated now take the updated data from bug table and 
+                    + " SELECT ApplicationName, BugSymptoms, BugTrigger, ProjectName, SourceFile, ClassName, LineNumber, Method, CodeBlock, CodeAuthor, SourceCode, Name, Date, Comment FROM BugTable WHERE BugID ='" + txtBugID.Text + "'", con);  // place it into the archive table then delete the bug form bug table. 
                 SqlCommand delete = new SqlCommand("DELETE FROM BugTable WHERE BugID = '" + bugID +"'" , con);
 
                 con.Open();
-                upadate.ExecuteNonQuery();
                 archive.ExecuteNonQuery();
                 delete.ExecuteNonQuery();
                 con.Close();
                 displayBugs();
-                MessageBox.Show("Bug ID: "+bugID + " Has been archive");
+                MessageBox.Show("Bug ID: "+bugID + " has been archive", "Alert", MessageBoxButtons.OK, MessageBoxIcon.Information);
             }
             else
             {
-                MessageBox.Show(" Please enter or select the bug ID you wish to archive!");
+                MessageBox.Show(" Please enter or select the bug ID you wish to archive and fill in all fields!", "Alert", MessageBoxButtons.OK, MessageBoxIcon.Error);
             }
 
         }
+        /// <summary>
+        /// Dsiplay the ArchivedBugs to the user
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
         private void btnViewArchivedBugs_Click(object sender, EventArgs e)
         {
             this.Hide();
@@ -191,11 +232,16 @@ namespace Debugger
                 e.Handled = true;
             }
         }
-
+        /// <summary>
+        ///  Display the users editable source code
+        ///  formatted and colour coded in a web browser
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
         private void btnEditSourceCode_Click(object sender, EventArgs e)
         {
-            String editedSourceCode = txtSourceCode.Text;
-            string colourSourceCode = new CodeColorizer().Colorize(editedSourceCode, Languages.CSharp);
+            String editedSourceCode = txtSourceCode.Text;                                                    
+            string colourSourceCode = new CodeColorizer().Colorize(editedSourceCode, Languages.CSharp);          // Format source code with colour for c# code with the colour code libary
             string html = ("<!doctype html><head><meta charset=\"utf-8\" <title> Code Snippet </title> </head> <body>" + colourSourceCode + "</body></html>");
             webColourCode.DocumentText = html;
